+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 ### 8.0.7
+* Added an "Ignore Readable" option to Destroy
 * Updated libraries